// src/task/constants/task-activity-event.enum.ts
export enum TaskActivityEvent {
  TASK_CREATED = 'TASK_CREATED',
  TASK_UPDATED = 'TASK_UPDATED',
  TASK_TITLE_CHANGED = 'TASK_TITLE_CHANGED',
  TASK_DESCRIPTION_CHANGED = 'TASK_DESCRIPTION_CHANGED',
  TASK_STATUS_CHANGED = 'TASK_STATUS_CHANGED',
  TASK_PRIORITY_CHANGED = 'TASK_PRIORITY_CHANGED',
  TASK_START_DATE_CHANGED = 'TASK_START_DATE_CHANGED',
  TASK_DUE_DATE_CHANGED = 'TASK_DUE_DATE_CHANGED',
  TASK_ESTIMATED_MINUTES_CHANGED = 'TASK_ESTIMATED_MINUTES_CHANGED',
  TASK_SPENT_MINUTES_CHANGED = 'TASK_SPENT_MINUTES_CHANGED',
  TASK_ASSIGNED = 'TASK_ASSIGNED',
  TASK_UNASSIGNED = 'TASK_UNASSIGNED',
  TASK_TAG_ADDED = 'TASK_TAG_ADDED',
  TASK_TAG_REMOVED = 'TASK_TAG_REMOVED',
  TASK_ATTACHMENT_ADDED = 'TASK_ATTACHMENT_ADDED',
  TASK_ATTACHMENT_REMOVED = 'TASK_ATTACHMENT_REMOVED',
  TASK_DELETED = 'TASK_DELETED',
  TASK_RESTORED = 'TASK_RESTORED',
}

export enum TaskActivityEntity {
  TASK = 'TASK',
  TASK_ASSIGNEE = 'TASK_ASSIGNEE',
  TASK_TAG = 'TASK_TAG',
  TASK_ATTACHMENT = 'TASK_ATTACHMENT',
}